Move `split_file` function into a `tools` directory
